AWS cloud support with terraform, containers
This was used for the recent cloud run (done on Sep 26, 2025).

See 60_cloud_helpers/README.md for detailed usage notes.

main.tf is the terraform script that allocates cloud resources.
(other examples are in tf-configs)

Apart from that there are a bunch of scripts which allocate and
track resources. Plus ferry data to the cloud, run the job on
the cloud and collect results.